requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiomysql
aiosqlite
httpx
aiohttp
beautifulsoup4
lxml
pydantic
pydantic[email]
pydantic-settings
python-dotenv
requests
jinja2
browser-cookie3
opencv-python
numpy
pytest-playwright
alembic
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart
asyncpg
metapub
DrissionPage